@@ -5,30 +5,51 @@
     "start": "webpack serve --mode development",
     "build": "webpack --mode production",
     "format": "prettier --write src",
-    "lint": "eslint src/**/*.js src/**/*.jsx"
-
+    "lint": "eslint --no-flat-config src/**/*.js src/**/*.jsx"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-brands-svg-icons": "^6.7.2",
+    "@fortawesome/free-regular-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@popperjs/core": "^2.11.8",
+    "@reduxjs/toolkit": "^2.8.2",
+    "axios": "^1.4.0",
+    "dayjs": "^1.11.13",
+    "primeicons": "^7.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^9.2.0",
     "react-router-dom": "^6.14.2",
-    "axios": "^1.4.0"
+    "react-toastify": "^11.0.5",
+    "reactstrap": "^9.2.3",
+    "redux": "^5.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.5",
-    "@babel/preset-env": "^7.22.5",
-    "@babel/preset-react": "^7.22.5",
-    "babel-loader": "^9.1.2",
-    "css-loader": "^6.7.3",
-    "eslint": "^8.44.0",
+    "@babel/core": "^7.27.4",
+    "@babel/eslint-parser": "^7.27.5",
+    "@babel/preset-env": "^7.27.2",
+    "@babel/preset-react": "^7.27.1",
+    "babel-loader": "^9.2.1",
+    "bootstrap": "^5.3.2",
+    "bootswatch": "^5.3.2",
+    "css-loader": "^6.11.0",
+    "eslint": "^9.29.0",
     "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-react": "^7.32.2",
-    "html-webpack-plugin": "^5.5.1",
+    "eslint-config-react": "^1.1.7",
+    "eslint-define-config": "^2.1.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-webpack-plugin": "^4.2.0",
+    "html-webpack-plugin": "^5.6.3",
     "prettier": "^3.0.0",
-    "style-loader": "^3.3.1",
-    "webpack": "^5.88.1",
+    "process": "^0.11.10",
+    "sass": "^1.66.1",
+    "sass-loader": "^13.3.3",
+    "style-loader": "^3.3.4",
+    "webpack": "^5.99.9",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1"
+    "webpack-dev-server": "^4.15.2"
   }
-  
 }
