@@ -1,8 +1,15 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import './styles/main.css';
+import React from 'react'
+import { createRoot } from 'react-dom/client'
+import App from './App'
+import { Provider } from 'react-redux'
+import store from './store'
+import { BrowserRouter } from 'react-router-dom'  
 
-const rootElement = document.getElementById('root');
-
-ReactDOM.render(<App />, rootElement);
+const root = createRoot(document.getElementById('root'))
+root.render(
+  <Provider store={store}>
+    <BrowserRouter>  
+      <App />
+    </BrowserRouter>
+  </Provider>
+)
