@@ -1,76 +1,55 @@
-import 'react-toastify/dist/ReactToastify.css';
-import './app.scss';
-import './theme.css';
-import './primereact.min.css';
-import 'primeicons/primeicons.css';
-
-import './config/dayjs';
 import React, { useEffect } from 'react';
-import { BrowserRouter as Router } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { getSession, logout } from './shared/reducers/auth-reducer.js';
-import { getProfile } from './shared/reducers/application-profile';
-import Header from './shared/layout/header/header';
+import authReducer, { login, logout } from './shared/reducers/auth.reducer.js';
+import Header from './shared/layout/header/header-component.jsx';
 import Footer from './shared/layout/footer/footer';
 import { hasAnyAuthority } from './shared/auth/private-route';
-import ErrorBoundary from './shared/error/error-boundary';
-import { AUTHORITIES } from './constants/constants';
-import AppRoutes from './routes';
+import AppRoutes from './routes.jsx';
 import useScrollToTopOnPageChange from './shared/hooks/useScrollToTop';
-import ToastContainerApp from './shared/componentes/toast-container.component';
+import ToastContainerApp from './shared/components/toast-container.component.jsx';
+import './app.scss';
 
 const baseHref = document.querySelector('base')?.getAttribute('href')?.replace(/\/$/, '') || '/';
 
 export const App = () => {
   const dispatch = useDispatch();
 
-
   useEffect(() => {
     const token = localStorage.getItem('token');
     if (token) {
       dispatch(getSession(token));
-      dispatch(getProfile());
     } else {
-      dispatch(logout()); 
+      dispatch(logout());
     }
   }, [dispatch]);
 
-  const currentLocale = useSelector(state => state.locale.currentLocale);
+  const currentLocale = useSelector(state => state.locale?.currentLocale || 'en');
   const { isAuthenticated, loading: isLoading, account } = useSelector(state => state.authentication);
-  const { ribbonEnv, inProduction: isInProduction, isOpenAPIEnabled } = useSelector(state => state.applicationProfile);
-  
-  const isAdmin = useSelector(state => 
-    hasAnyAuthority(state.authentication.account?.authorities || [], [AUTHORITIES.ADMIN])
+
+  const isAdmin = useSelector(state =>
+    hasAnyAuthority(state.authentication.account?.authorities)
   );
 
   useScrollToTopOnPageChange();
 
   return (
-    <Router basename={baseHref}>
-      <div className="app-container d-flex flex-column">
-        <ToastContainerApp />
+    <div className="app-container d-flex flex-column min-vh-100">
+      <ToastContainerApp />
+      <Header
+        isLoading={isLoading}
+        isAuthenticated={isAuthenticated}
+        isAdmin={isAdmin}
+        currentLocale={currentLocale}
+        nombreDeUsuario={isAuthenticated ? account.login : ''}
+      />
 
-        <ErrorBoundary>
-          <Header
-            isLoading={isLoading}
-            isAuthenticated={isAuthenticated}
-            isAdmin={isAdmin}
-            currentLocale={currentLocale}
-            ribbonEnv={ribbonEnv}
-            isInProduction={isInProduction}
-            isOpenAPIEnabled={isOpenAPIEnabled}
-            nombreDeUsuario={isAuthenticated ? account.login : ''} 
-          />
-        </ErrorBoundary>
+      <main className="flex-grow-1">
+        <AppRoutes />
+      </main>
 
-        <ErrorBoundary>
-          <AppRoutes />
-        </ErrorBoundary>
-
-        <Footer isAuthenticated={isAuthenticated} />
-      </div>
-    </Router>
+      <Footer isAuthenticated={isAuthenticated} />
+    </div>
   );
 };
 
