@@ -1,25 +1,17 @@
 import React from 'react';
-import { Translate } from 'react-jhipster';
-import { NavItem, NavLink } from 'reactstrap';
-import { Link } from 'react-router-dom';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import ability from '../config/ability';
+import { Nav, Navbar } from 'reactstrap';
+import { Brand, Home } from './header-component'; 
 
-export const Home = () => {
-  if (!ability.can('access', 'menu-inicio')) {
-    return null;
-  }
-
+const Header = ({ isAuthenticated, isLoading, /* otros props */ }) => {
   return (
-    <NavItem>
-      <NavLink tag={Link} to="/" className="d-flex align-items-center">
-        <FontAwesomeIcon icon="home" />
-        <span className="ml-2">
-          <Translate contentKey="global.menu.home">Inicio</Translate>
-        </span>
-      </NavLink>
-    </NavItem>
+    <Navbar color="light" light expand="md">
+      <Brand />
+      <Nav navbar>
+        <Home /> 
+       
+      </Nav>
+    </Navbar>
   );
 };
 
-export default Home;
+export default Header;
